SLCORE-974 Load the appropriate embedded C# analyzer
